chromatogramandspectra to modular version

+++ src/main/java/io/github/mzmine/modules/visualization/chromatogramandspectra/SpectraDataSetCalc.java
@@ -69,13 +69,13 @@ public class SpectraDataSetCalc extends AbstractTask {
     setStatus(TaskStatus.PROCESSING);
 
     if (showSpectraOfEveryRawFile) {
-      float rt = pos.getRetentionTime();
+      double rt = pos.getRetentionTime();
       rawDataFiles.forEach(rawDataFile -> {
         int num = -1;
         if (scanSelection.getMsLevel() != null) {
-          num = rawDataFile.getScanNumberAtRT(rt, scanSelection.getMsLevel());
+          num = rawDataFile.getScanNumberAtRT((float) rt, scanSelection.getMsLevel());
         } else {
-          num = rawDataFile.getScanNumberAtRT(rt);
+          num = rawDataFile.getScanNumberAtRT((float) rt);
         }
         if (num != -1) {
           Scan scan = rawDataFile.getScan(num);
